Remove old code and rename *set to *
