+++ packages/bundle/src/boot/actual/component/minimal.ts
