+++ packages/bundle/src/boot/actual/hook/full.ts
@@ -1,4 +1,4 @@
-export * from './minimalSet';
+export * from './minimal';
 
 // Adds.
 export { default as useAdaptiveCardsHostConfig } from '../../../adaptiveCards/hooks/useAdaptiveCardsHostConfig';
