--- packages/bundle/src/boot/actual/hook/fullDiff.ts
@@ -1,5 +1,0 @@
-// // Additional
-// export { default as useAdaptiveCardsHostConfig } from '../../../adaptiveCards/hooks/useAdaptiveCardsHostConfig';
-// export { default as useAdaptiveCardsPackage } from '../../../adaptiveCards/hooks/useAdaptiveCardsPackage';
-// export { default as useStyleOptions } from '../../../hooks/useStyleOptions';
-// export { default as useStyleSet } from '../../../hooks/useStyleSet';
