+++ packages/bundle/src/boot/actual/hook/minimal.ts
