+++ packages/bundle/src/boot/actual/minimal.ts
@@ -9,9 +9,9 @@ const { object: buildInfoObject, version } = buildInfo;
 export { concatMiddleware, createStyleSet, testIds, withEmoji } from 'botframework-webchat-component';
 export { Constants, createStore, createStoreWithDevTools, createStoreWithOptions } from 'botframework-webchat-core';
 export { default as createBrowserWebSpeechPonyfillFactory } from '../../createBrowserWebSpeechPonyfillFactory';
-export { ReactWebChat as default, ReactWebChat } from './component/fullSet';
+export { ReactWebChat as default, ReactWebChat } from './component/full';
 export { default as createDirectLine } from './private/createDirectLine';
 export { default as createDirectLineAppServiceExtension } from './private/createDirectLineAppServiceExtension';
-export { renderWebChat } from './renderWebChat/fullSet';
-export { type StrictStyleOptions, type StyleOptions } from './styleOptions/minimalSet';
+export { renderWebChat } from './renderWebChat/full';
+export { type StrictStyleOptions, type StyleOptions } from './styleOptions/minimal';
 export { buildInfoObject as buildInfo, version };
