+++ packages/bundle/src/boot/actual/reactWebChat/full.ts
@@ -1,0 +1,1 @@
+export { ReactWebChat } from '../component/full';
