--- packages/bundle/src/boot/actual/reactWebChat/fullSet.ts
@@ -1,1 +1,0 @@
-export { ReactWebChat } from '../component/fullSet';
