+++ packages/bundle/src/boot/actual/reactWebChat/minimal.ts
@@ -1,0 +1,1 @@
+export { ReactWebChat } from '../component/minimal';
