--- packages/bundle/src/boot/actual/reactWebChat/minimalSet.ts
@@ -1,1 +1,0 @@
-export { ReactWebChat } from '../component/minimalSet';
