+++ packages/bundle/src/boot/actual/renderWebChat/full.ts
@@ -1,5 +1,5 @@
 import coreRenderWebChat from '../../../renderWebChat';
-import { ReactWebChat } from '../reactWebChat/minimalSet';
+import { ReactWebChat } from '../reactWebChat/full';
 
 const renderWebChat = coreRenderWebChat.bind(null, ReactWebChat);
