--- packages/bundle/src/boot/exports/component.old.ts
@@ -1,5 +1,0 @@
-// export * from '../actual/component/minimalSet';
-// export * from '../actual/component/fullDiff';
-
-// // Explicit overrides.
-// export { Composer } from '../actual/component/fullDiff';
