--- packages/bundle/src/boot/exports/hook.old.ts
@@ -1,5 +1,0 @@
-// export * from '../actual/hook/minimalSet';
-// export * from '../actual/hook/fullOnly';
-
-// // Explicit overrides.
-// export { useStyleOptions, useStyleSet } from '../actual/hook/fullOnly';
