+++ packages/bundle/src/boot/exports/index.ts
@@ -1,14 +1,21 @@
-export * from '../actual/minimalSet';
+export * from '../actual/minimal';
 
 // #region Overrides
-export * from '../actual/fullSet';
-export { createStyleSet, version, type StrictStyleOptions, type StyleOptions } from '../actual/fullSet';
+export * from '../actual/full';
+export {
+  createStyleSet,
+  ReactWebChat,
+  renderWebChat,
+  version,
+  type StrictStyleOptions,
+  type StyleOptions
+} from '../actual/full';
 // #endregion
 
 // #region Backward compatibility
 // Web devs should move to named exports instead.
+export * as Components from '../actual/component/full';
 export * as decorator from '../actual/decorator';
+export * as hooks from '../actual/hook/full';
 export * as internal from '../actual/internal';
-export * as Components from '../actual/component/fullSet';
-export * as hooks from '../actual/hook/fullSet';
 // #endregion
