+++ packages/bundle/src/boot/exports/minimal.ts
@@ -1,9 +1,9 @@
-export * from '../actual/minimalSet';
+export * from '../actual/minimal';
 
 // #region Backward compatibility
 // Web devs should move to named exports instead.
 export * as decorator from '../actual/decorator';
 export * as internal from '../actual/internal';
-export * as Components from '../actual/component/minimalSet';
-export * as hooks from '../actual/hook/minimalSet';
+export * as Components from '../actual/component/minimal';
+export * as hooks from '../actual/hook/minimal';
 // #endregion
